+++ example_runner.py
@@ -1,0 +1,36 @@
+print('building examples')
+import example.base_coffin
+import example.frame_window
+import example.frame
+import example.ramp_frame
+import example.ramp
+import example.rampGreebled
+import example.rotation_demo
+
+print('building container')
+import example.container.container_greeble_hack
+import example.container.container_tile
+import example.container.container
+import example.container.containerDoor
+
+print('building portal')
+import example.portal.portal_greebled_ramp_window
+import example.portal.portal_greebled_ramp
+import example.portal.portal
+import example.portal.portalBase
+
+print('building shieldwall')
+import example.shieldwall.archset
+import example.shieldwall.corner_connector
+import example.shieldwall.curve_basic
+import example.shieldwall.endcap
+import example.shieldwall.gothicMesh
+import example.shieldwall.hexMesh
+import example.shieldwall.hexset
+import example.shieldwall.magnets
+import example.shieldwall.mesh
+import example.shieldwall.set
+import example.shieldwall.shape
+import example.shieldwall.straight_basic
+import example.shieldwall.straight_hex
+import example.shieldwall.straight